requirements.txt
django>=2.1.6
django-debug-toolbar==1.11
django-debug-toolbar-request-history==0.0.9
et-xmlfile==1.0.1
jdcal==1.4
lml==0.0.9
openpyxl==2.6.1
pyexcel==0.5.12
pyexcel-io==0.5.14
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.7
pytz==2018.9
sqlparse==0.2.4
texttable==1.6.1
xlrd==1.2.0
xlwt==1.3.0
djangorestframework
django-cors-headers